fix: fix content file renaming
