+++ src/index.tsx
@@ -8,7 +8,7 @@ import { QueryClientProvider } from '@tanstack/react-query'
 import contextMenuPatch, { LibraryContextMenu } from './patches/context-menu'
 import { getSettingsQueryOptions, queryClient } from './query'
 import { libraryAppPatch } from './patches/library-app'
-import { Content } from "./components/Content"
+import { Content } from "./components/content"
 import { batchCreateLaunchOptions } from "./components/batch-add-launch-options"
 
 export default definePlugin(() => {
